Issue #16: Refactored folder structure.
 - Moved low level functionality to 'infrastructure'.
 - Changed usecases to package with boundaries in __init__.py and usecases are now capsulated into classes.
 - TeamSpeak users are updated via usecases instead of relying to lower level user model.
