+++ tessumod/src/scripts/client/gui/mods/tessumod/usecases/entities.py
@@ -15,37 +15,47 @@
 # License along with this library; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 
-class PositionalData(object):
-
-	def __init__(self):
-		self.camera_position = None
-		self.camera_direction = None
-		self.client_positions = {}
-
-class Vector(object):
-
-	def __init__(self, x, y, z):
-		self.x = x
-		self.y = y
-		self.z = z
-
 class TeamSpeakUser(object):
 
-	def __init__(self):
-		self.nick = None
-		self.wot_nick = None
-		self.client_id = None
-		self.unique_id = None
-		self.channel_id = None
-		self.speaking = False
+	def __init__(self, nick, game_nick, client_id, unique_id, channel_id, speaking):
+		self.__nick = nick
+		self.__game_nick = game_nick
+		self.__client_id = client_id
+		self.__unique_id = unique_id
+		self.__channel_id = channel_id
+		self.__speaking = speaking
+
+	@property
+	def nick(self):
+		return self.__nick
+
+	@property
+	def game_nick(self):
+		return self.__game_nick
+
+	@property
+	def client_id(self):
+		return self.__client_id
+
+	@property
+	def unique_id(self):
+		return self.__unique_id
+
+	@property
+	def channel_id(self):
+		return self.__channel_id
+
+	@property
+	def speaking(self):
+		return self.__speaking
 
 	def __hash__(self):
 		return (
-			hash(self.client_id) ^
-			hash(self.nick) ^
-			hash(self.wot_nick) ^
-			hash(self.unique_id) ^
-			hash(self.channel_id)
+			hash(self.__client_id) ^
+			hash(self.__nick) ^
+			hash(self.__game_nick) ^
+			hash(self.__unique_id) ^
+			hash(self.__channel_id)
 		)
 
 	def __eq__(self, other):
@@ -53,12 +63,12 @@ class TeamSpeakUser(object):
 
 	def __repr__(self):
 		return "TeamSpeakUser(client_id={0}, nick={1}, wot_nick={2}, unique_id={3}, channel_id={4}, speaking={5})".format(
-			repr(self.client_id),
-			repr(self.nick),
-			repr(self.wot_nick),
-			repr(self.unique_id),
-			repr(self.channel_id),
-			repr(self.speaking)
+			repr(self.__client_id),
+			repr(self.__nick),
+			repr(self.__game_nick),
+			repr(self.__unique_id),
+			repr(self.__channel_id),
+			repr(self.__speaking)
 		)
 
 class GamePlayer(object):
